Removed next-auth, instead used cookie and jose.
Next auth was causing too many problems trying to stay logged in. In my debugging, i downgraded the react version out of the beta. The problem is the next-auth is just flat out broken when trying to add custom cookies when reloading a page.
To view a User, import UserContext. the data for the context can be found in lib/setCookie.
.env needs these:  NEXTAUTH_URL="http://localhost:3000/"
JWT_SECRET="random secret key"